main.py
import yaml
import torch
from preprocessing.graph_builder import GraphBuilder
from models.gnn_models import GCNBot, GraphSAGEBot, GATBot
from training.trainer import BotDetectionTrainer
from training.evaluator import ModelEvaluator

def load_config(config_path='configs/config.yaml'):
    """Load configuration file"""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config

def main():
    """Main execution function"""
    print("=" * 50)
    print("NeuroShield: GNN-based Bot Detection")
    print("=" * 50)
    
    # Load config
    config = load_config()
    print("\n✓ Configuration loaded")
    
    # TODO: Load and process data
    print("\n[1/5] Loading data...")
    # graph_builder = GraphBuilder()
    # data = graph_builder.load_data(...)
    
    # TODO: Initialize model
    print("\n[2/5] Initializing model...")
    # model = GCNBot(...)
    
    # TODO: Train model
    print("\n[3/5] Training model...")
    # trainer = BotDetectionTrainer(model, data, config)
    # trainer.train(num_epochs=100)
    
    # TODO: Evaluate model
    print("\n[4/5] Evaluating model...")
    # evaluator = ModelEvaluator(model, data)
    # evaluator.print_report()
    
    # TODO: Save model
    print("\n[5/5] Saving model...")
    # torch.save(model.state_dict(), 'saved_models/neuroshield_model.pt')
    
    print("\n✓ Training complete!")

if __name__ == "__main__":
    main()
