requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web app
streamlit>=1.28.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: Explainability
shap>=0.42.0
captum>=0.6.0
